backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langgraph
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pytest
httpx
langchain-groq